chore(utils): consolidate split-position helpers into shared/utils
getSplitPositions, getSplitPieceCount, and getSplitPlanePositionsMm
lived in src/features/bin-designer/utils/ with a near-identical copy
of getSplitPositions inlined in grid-editor's SplitLineOverlay. The
grid-editor copy still used the old greedy-halving algorithm (pre
#1400 fix) — it only didn't manifest the bug because its callers
always pass integer grid sizes.

Move the three helpers to src/shared/utils/ so both features share
the same implementation and future fixes apply once. Also relocates
the orphaned computePinPositions tests to sit alongside the function
they test in src/shared/generation/.

No behavior change for bin-designer; grid-editor's isometric preview
now uses the minimum-equal-pieces algorithm, which produces the same
split lines as before for the integer grid sizes it's given.
