+++ src/features/bin-designer/hooks/useSplitPreview.ts
@@ -19,10 +19,7 @@ import { useDesignerStore } from '@/features/bin-designer/store';
 import { useSettingsStore } from '@/core/store/settings';
 import { calcMaxGridUnits } from '@/core/constants';
 import { getActiveBridge, workerPoolManager } from '@/shared/generation/bridge';
-import {
-  getSplitPieceCount,
-  getSplitPlanePositionsMm,
-} from '@/features/bin-designer/utils/splitPositions';
+import { getSplitPieceCount, getSplitPlanePositionsMm } from '@/shared/utils/splitPositions';
 import { DEFAULT_SPLIT_CONNECTOR_CONFIG } from '@/features/bin-designer/constants/defaults';
 import type { SplitPieceMeshEntry } from '../types';
