+++ src/features/bin-designer/utils/index.ts
@@ -27,7 +27,6 @@ export {
   setPreviewCanvas,
   clearPreviewCanvas,
 } from './thumbnail';
-export { getSplitPositions, getSplitPieceCount, getSplitPlanePositionsMm } from './splitPositions';
 export { packageSplitPiecesAsZip } from './splitExport';
 export { validateBinParams, computeMinCellSize, validateCompartmentSizes } from './validation';
 export type { DesignerValidationError, MinCellSize } from './validation';
